FX-689 (preliminary): added subtree/parent tree locking.
This solution is still not satisfying, because InnoDB acquires
locks for every row it touches in an UPDATE or 
SELECT ... FOR UPDATE. Since the 'total children'
update queries are specified by a range, they unavoidably
touch rows that shouldn't be locked, thus increasing
the potential for deadlocks. Either the range queries
must be performed through a direct index lookup (e.g. by ID),
or they should be disabled (at least for MySQL),
or we need a generic proxy for dealing with deadlocks.



git-svn-id: http://svn.flexive.org/repository/flexive/trunk@1688 d35d4baa-7912-0410-b0ae-edacad209e15