FX-689 (preliminary): remove TOTAL_CHILDCOUNT computation from getNode
The TOTAL_CHILDCOUNT computation in getNode() is problematic,
because it usually involves a full table
scan in MySQL (even though the left and right columns
are indexed). This can lead to similar issues since the
affected node IDs are not specified explicitly, although
in this case only a shared lock is requested.

FX-690 suggests the addition of an TreeEngine API method to
calculate the total child count instead.

git-svn-id: http://svn.flexive.org/repository/flexive/trunk@1702 d35d4baa-7912-0410-b0ae-edacad209e15