* Fix for 'delete' property permissions
Broke out the 'delete' check in its own loop, since it's
fundamentally different to the other permission checks.

Rejecting permission based on the first non-allowed ACL 
*assignment* is wrong because an ACL may be assigned 
multiple times (once for each group the user is in). This
works for multiple content ACLs only because the logic is inverted
(allowing permission when at least one assignment is valid). 



git-svn-id: http://svn.flexive.org/repository/flexive/trunk@1814 d35d4baa-7912-0410-b0ae-edacad209e15