JSF2: removed fx:yuiSetup, specify all dependencies on components
The call to fx:yuiSetup is not necessary in JSF2 applications, since we can
just include a initializer JavaScript (js/init.js) at the end of the body instead.

However, fx:includes is still necessary because it contains expressions that
need to be evaluated server-side at runtime that cannot be guessed from a
pure client-side JavaScript (like the user and GUI languages). 



git-svn-id: http://svn.flexive.org/repository/flexive/trunk@2336 d35d4baa-7912-0410-b0ae-edacad209e15