FxSQL: added hint for skipping the result set information (total row count, found content types)
When supported by the DBMS, this can lead to improved performance for queries
that don't need to display these informations to the user or don't use paging,
since we don't need to collect the results in a temporary result table.  

The generated SQL is not optimal since the source table for the data selector 
is simply replaced with the original select, but if the DBMS optimizes the query
well enough, it does not seem to impact performance too much.

git-svn-id: http://svn.flexive.org/repository/flexive/trunk@2726 d35d4baa-7912-0410-b0ae-edacad209e15