PostgreSQL sequencer storage: fix race condition in getCustomSequencers
Since each sequence is a relation and sequence removal means deleting that relation, this change will be observed by other threads even when using a transaction. Reading the sequencer relation data in a non-transactional connection prevents unrecoverable errors in case a sequence is deleted in a different thread while getCustomSequencers is being run.

git-svn-id: http://svn.flexive.org/repository/flexive/trunk@2791 d35d4baa-7912-0410-b0ae-edacad209e15