Content storage: drop group positions from FX_CONTENT_DATA
Depending on the data structure, this can remove up to 30-40% of rows in FX_CONTENT_DATA, at the expense of a single CLOB column in the main FX_CONTENT table per instance version. Existing databases will be migrated on startup, depending on the database size this may take several minutes.

git-svn-id: http://svn.flexive.org/repository/flexive/trunk@2864 d35d4baa-7912-0410-b0ae-edacad209e15